Fixes GET request on undefined image
The UI re-render's as soon as `this.photo` is set, as it's  annotaed with `@State`.

The problem comes up when the UI renders faster as as `getOrientation` and the following setting of `this.photoSrc` can execute (which happens reproducible for me).

With the UI already rendering and `photoSrc` not beeing set, the `accept-image` tries to request `undefined`.

To fix this, I moved setting `this.photo` to the end of  `promptAccept`.

I did the following additional changes:
* Removed `@State` from `photoSrc` to save one redundant render cycle (and be consistent with `photoOrientation`)
* Removed various `console.log` to keep the console clean in production apps.